Suppress Expect 100-continue header
By default, curl sends a header with the Expect: 100-continue field for
large POST data; this is inconvenient for some servers which reject such
expectations. In keeping with hurl's philosophy of "make doing the wrong
thing easy," we just override this behavior and force curl to send all
the data with the initial request.

If this approach causes problems, we will switch to doing a different
wrong thing.